backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
requests==2.32.3
geopy==2.4.1
Pillow==11.0.0
numpy==2.2.1
opencv-python-headless==4.10.0.84
scikit-image==0.25.0
httpx==0.28.1
python-dotenv==1.0.1
email-validator==2.2.0
aiofiles==24.1.0
psutil==5.9.8
psycopg2-binary==2.9.10
